refactor: move Plugin and Theme models to App\Models\WpOrg
