+++ app/Models/WpOrg/Plugin.php
@@ -1,11 +1,12 @@
 <?php
 
-namespace App\Models;
+namespace App\Models\WpOrg;
 
+use App\Models\BaseModel;
 use Illuminate\Database\Eloquent\Concerns\HasUuids;
 
 /**
- * 
+ *
  *
  * @method static \Illuminate\Database\Eloquent\Builder<static>|Plugin newModelQuery()
  * @method static \Illuminate\Database\Eloquent\Builder<static>|Plugin newQuery()
